Get backdrop blur working on hover
For some reason, it doesn't work if the blur and background opacity are on the same object - no idea why though. So, I split it into 2 divs.

Note that I also tried using a pseudo-element to avoid the extra divs, but they apparently don't work on img tags, so that didn't pan out.